+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### 🐞 Bugs Fixed
 - Analysis result information not shown when CSRF Protection is turned off (fixes [#73](https://github.com/jenkinsci/dependency-track-plugin/issues/73))
+- The threshold for new findings used the last build, even though it may not have had a Dependency-Track analysis result.
 
 ## v4.0.0 - 2021-08-29
 ### ⚠ Breaking